[SmarterMail100] Fix for large numbers when going from MB to kb
Tested with SmarterMail Free 100.0.9088.25586 (Nov 18, 2024)

+++ SolidCP/Sources/SolidCP.Providers.Mail.SmarterMail100/SmarterMail100.cs
@@ -818,7 +818,7 @@ namespace SolidCP.Providers.Mail
 					userName = mailbox.Name,
 					fullName = mailbox.FullName,
 					password = mailbox.Password,
-					maxMailboxSize = mailbox.MaxMailboxSize * 1048576
+					maxMailboxSize = (long)mailbox.MaxMailboxSize * 1048576
                 };
 
 				var userContactInfoArray = new
